transaction/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
mbbank-lib
httpx==0.27.0
httpx==0.27.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
